Version 1.1.0, algorithm changes and bugfixes
Handle a empty note vectors when logging number of notes and time. Account for tempo when
reporting time.

Log the application version.

weight_discord expects a voice number as its 2nd argument but it was getting a pitch.
Self-dissonance was not being ignored.

Divide the delta-t for new notes by 4 instead of the number of voices to avoid time
intervals that aren't multiples of 16th notes. This helps with importing MIDI files into
other programs.

Changed the point of interest to be one note earlier, so it can be described relatively
succinctly as an index when consonance peaks after crossing the midpoint.

Replaced old examples with new ones.