Added the ability to get defaults from a log file
If a .log file is given as an argument, settings are taken from that file. Command-line
options override those options.

There was a subtle bug in development. If the key was random when the log file was
generated, it was being set explicitly from the value in the log file. Since a random
number wasn't generated, the generator was in a different state when it was used for note
generation. If an option other than seed is random in the log file it must also be generated
when the log file is passed as an argument. The seed guarantees it will be the same value
for the option.